Tech debt/refactor

+++ service/static/js/rest_api.js
@@ -38,6 +38,11 @@ $(function () {
         let product_b_sku = $("#recommendation_product_b_sku").val();
         let recommendation_type = $("#recommendation_recommendation_type").val();
 
+        if (!product_a_sku || !product_b_sku || !recommendation_type) {
+            flash_message("Please complete the attributes of the Recommendation: product_a_sku, product_b_sku, recommendation_type.");
+            return;
+        }
+
         let data = {
             "product_a_sku": product_a_sku,
             "product_b_sku": product_b_sku,
@@ -48,7 +53,7 @@ $(function () {
 
         let ajax = $.ajax({
             type: "POST",
-            url: "/recommendations",
+            url: "/api/recommendations",
             contentType: "application/json",
             data: JSON.stringify(data),
         });
@@ -75,6 +80,11 @@ $(function () {
         let product_b_sku = $("#recommendation_product_b_sku").val();
         let recommendation_type = $("#recommendation_recommendation_type").val();
 
+        if (!recommendation_id || !product_a_sku || !product_b_sku || !recommendation_type) {
+            flash_message("Please complete the attributes of the Recommendation: id, product_a_sku, product_b_sku, recommendation_type.");
+            return;
+        }
+
         let data = {
             "product_a_sku": product_a_sku,
             "product_b_sku": product_b_sku,
@@ -85,7 +95,7 @@ $(function () {
 
         let ajax = $.ajax({
             type: "PUT",
-            url: `/recommendations/${recommendation_id}`,
+            url: `/api/recommendations/${recommendation_id}`,
             contentType: "application/json",
             data: JSON.stringify(data)
         })
@@ -109,11 +119,16 @@ $(function () {
 
         let recommendation_id = $("#recommendation_id").val();
 
+        if (!recommendation_id) {
+            flash_message("Please select a recommendation to retrieve.");
+            return;
+        }
+
         $("#flash_message").empty();
 
         let ajax = $.ajax({
             type: "GET",
-            url: `/recommendations/${recommendation_id}`,
+            url: `/api/recommendations/${recommendation_id}`,
             contentType: "application/json",
             data: ''
         })
@@ -139,11 +154,16 @@ $(function () {
 
         let recommendation_id = $("#recommendation_id").val();
 
+        if (!recommendation_id) {
+            flash_message("Please select a recommendation to delete.");
+            return;
+        }
+
         $("#flash_message").empty();
 
         let ajax = $.ajax({
             type: "DELETE",
-            url: `/recommendations/${recommendation_id}`,
+            url: `/api/recommendations/${recommendation_id}`,
             contentType: "application/json",
             data: '',
         })
@@ -194,7 +214,7 @@ $(function () {
 
         let ajax = $.ajax({
             type: "GET",
-            url: `/recommendations?${queryString}`,
+            url: `/api/recommendations?${queryString}`,
             contentType: "application/json",
             data: ''
         })
@@ -244,7 +264,7 @@ $(function () {
 
         let ajax = $.ajax({
             type: "GET",
-            url: "/recommendations",
+            url: "/api/recommendations",
             contentType: "application/json",
             data: ''
         })
@@ -293,20 +313,20 @@ $(function () {
             flash_message("Please select a recommendation to like.");
             return;
         }
-    
+
         $("#flash_message").empty();
-    
+
         let ajax = $.ajax({
             type: "PUT",
-            url: `/recommendations/${recommendation_id}/like`,
+            url: `/api/recommendations/${recommendation_id}/like`,
             contentType: "application/json"
         });
-    
+
         ajax.done(function (res) {
             update_form_data(res);
             flash_message("Successfully liked the recommendation!");
         });
-    
+
         ajax.fail(function (res) {
             flash_message(res.responseJSON.message);
         });
@@ -322,20 +342,20 @@ $(function () {
             flash_message("Please select a recommendation to dislike.");
             return;
         }
-    
+
         $("#flash_message").empty();
-    
+
         let ajax = $.ajax({
             type: "DELETE",
-            url: `/recommendations/${recommendation_id}/like`,
+            url: `/api/recommendations/${recommendation_id}/like`,
             contentType: "application/json"
         });
-    
+
         ajax.done(function (res) {
             update_form_data(res);
             flash_message("Successfully disliked the recommendation!");
         });
-    
+
         ajax.fail(function (res) {
             flash_message(res.responseJSON.message);
         });
